feat!(nix): Migrate To Layered Approach
Migrate to layered approach to allow for using the Nix flake across
machines independent of OS and chip. This change also supports systems
which do not support Nix, such as locked down work computers. Add new
scripts which will help with the installation process to help keep
systems in sync. The prefered method will always be Nix. However, we can
now use Stow and manual linking for situations where Nix is not allowed
or possible.